refactor(llm): move models
