+++ src/a2a_t/llm/base.py
@@ -3,44 +3,13 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from dataclasses import dataclass, field
 from datetime import UTC, datetime
 from typing import Any
 from uuid import uuid4
 
 from a2a_t.llm.errors import LLMConfigError, LLMRuntimeError
+from a2a_t.llm.models import ChatMessage, ChatSession, LLMResponse
 from a2a_t.llm.session_store import InMemorySessionStore, ProviderScopedSessionStore, SessionStore
-
-
-@dataclass
-class ChatMessage:
-    """Represent one message in a chat session."""
-
-    role: str
-    content: str
-
-
-@dataclass
-class ChatSession:
-    """Persist chat history and metadata for a provider-scoped session."""
-
-    session_id: str
-    provider: str
-    system_prompt: str | None = None
-    messages: list[ChatMessage] = field(default_factory=list)
-    created_at: datetime = field(default_factory=lambda: datetime.now(UTC))
-    last_accessed_time: datetime = field(default_factory=lambda: datetime.now(UTC))
-
-
-@dataclass
-class LLMResponse:
-    """Response from an LLM adapter."""
-
-    content: str
-    model: str
-    usage: dict[str, int]
-    metadata: dict[str, Any]
-    session_id: str | None = None
 
 
 class LLMAdapter(ABC):
