+++ tests/prompt_compliance/test_integration_runtime.py
@@ -17,7 +17,7 @@ if str(SRC_ROOT) not in sys.path:
 from a2a_t.common.prompt_resources import PromptResourceLoader, SlotSchemaLoader, TemplateLoader
 from a2a_t.common.prompt_resources.models import ScenarioDefinition
 from a2a_t.common.prompt_resources.slot_json_schema_loader import SlotJsonSchemaLoader
-from a2a_t.llm.base import LLMResponse
+from a2a_t.llm.models import LLMResponse
 from a2a_t.prompt.analysis import SlotExtractor
 from a2a_t.prompt.analysis.models import ScenarioResolutionResult
 from a2a_t.prompt.common.models import PromptReference
