+++ tests/test_llm/test_base_chat_flow.py
@@ -11,7 +11,8 @@ SRC_ROOT = PROJECT_ROOT / "src"
 if str(SRC_ROOT) not in sys.path:
     sys.path.insert(0, str(SRC_ROOT))
 
-from a2a_t.llm.base import ChatMessage, LLMAdapter, LLMResponse
+from a2a_t.llm.base import LLMAdapter
+from a2a_t.llm.models import ChatMessage, LLMResponse
 from a2a_t.llm import ChatMessage as PublicChatMessage, ChatSession, InMemorySessionStore, LLMAdapterFactory, LLMClient
 from a2a_t.llm.errors import LLMRuntimeError
