+++ tests/test_llm/test_client.py
@@ -12,7 +12,7 @@ SRC_ROOT = PROJECT_ROOT / "src"
 if str(SRC_ROOT) not in sys.path:
     sys.path.insert(0, str(SRC_ROOT))
 
-from a2a_t.llm.base import ChatMessage, ChatSession, LLMResponse
+from a2a_t.llm.models import ChatMessage, ChatSession, LLMResponse
 from a2a_t.llm.errors import LLMConfigError, LLMRuntimeError
 from tests.test_support import ManagedTempDirTestCase
